refactor(es): Extract client construction into clientbuilder package
Decouple the config package from heavy ES client dependencies by moving
NewClient() and all related helpers (auth, transport, bulk processing)
into a new internal/storage/elasticsearch/clientbuilder package.

The config package is now a pure data package with no imports of
olivere/elastic, go-elasticsearch, or other runtime dependencies.

Config.Version changed from es.BackendVersion to uint so that the
config layer doesn't depend on the ES abstraction types. The builder
converts uint→BackendVersion internally after server detection.

Closes #8841
