+++ internal/storage/metricstore/elasticsearch/factory.go
@@ -9,6 +9,7 @@ import (
 	"go.opentelemetry.io/collector/extension/extensionauth"
 
 	es "github.com/jaegertracing/jaeger/internal/storage/elasticsearch"
+	"github.com/jaegertracing/jaeger/internal/storage/elasticsearch/clientbuilder"
 	"github.com/jaegertracing/jaeger/internal/storage/elasticsearch/config"
 	"github.com/jaegertracing/jaeger/internal/storage/elasticsearch/indices"
 	"github.com/jaegertracing/jaeger/internal/storage/v1/api/metricstore"
@@ -33,7 +34,7 @@ func NewFactory(
 		return nil, err
 	}
 
-	client, err := config.NewClient(ctx, &cfg, telset.Logger, telset.Metrics, httpAuth)
+	client, err := clientbuilder.NewClient(ctx, &cfg, telset.Logger, telset.Metrics, httpAuth)
 	if err != nil {
 		return nil, err
 	}
